Add dropdown in sidebar
