+++ src/components/SideBar/index.css
@@ -38,6 +38,7 @@
   margin: 0 0 10px;
   border-radius: 10px;
   transition: all 0.5s ease;
+  position: relative;
 }
 
 #sidebar-home-link {
@@ -74,6 +75,51 @@
   height: 30px;
   width: 30px;
   border-radius: 100%;
+  border: 1px solid #eee;
+}
+
+.sidebar_user_dropdown {
+  display: flex;
+}
+
+.sidebar_user_dropdown_container {
+  position: absolute;
+  top: 390px;
+  left: 100px;
+  padding: 0;
+  border-radius: 10px;
+  width: 170px;
+}
+
+.sidebar_user_dropdown_sub_container {
+  border: 1px solid grey;
+  background-color: var(--dark-post-bg);
+  color: var(--color);
+  padding-inline: 0px;
+  list-style-type: none;
+  border-radius: inherit;
+  box-shadow: var(--post-box-shadow);
+}
+
+.sidebar_user_dropdown_sub_container > li {
+  padding: 10px 25px;
+  margin: 0;
+  border-radius: 8px;
+  border-bottom: 1px solid #eee;
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  cursor: pointer;
+}
+
+.sidebar_user_dropdown_sub_container > li:hover {
+  background: var(--btn-color-hover);
+}
+.dropdown-list-profile-picture {
+  width: 25px;
+  border-radius: 50%;
+  border: 1px solid #eee;
+  margin-left: -3px;
 }
 
 footer {
@@ -145,6 +191,15 @@ footer {
     display: none;
   }
 
+  .sidebar_user_dropdown_container {
+    width: 150px;
+    left: 40px;
+  }
+
+  .sidebar_user_dropdown_sub_container > li {
+    padding-right: 0;
+  }
+
   footer {
     font-size: 10px;
   }
@@ -178,6 +233,12 @@ footer {
 
   .sidebar-links li {
     margin: auto;
+  }
+
+  .sidebar_user_dropdown_container {
+    left: initial;
+    top: -120px;
+    right: 5px;
   }
 
   footer {
